Move message factory methods to separate classes (common, money).
